Rename writeManifests
