+++ src/files/manifest/writeManifest.ts
